requirements.txt
pillow
pandas
openpyxl
requests
